Fix encoding issue
When starting a new ProcessBuilder process it does not load the defaultCharset (example: windows-1252), instead it loads some other charset. (example: ibm850). Read more about this issue here: http://stackoverflow.com/questions/22349139/utf-8-output-from-powershell/22350989
In the jPowerShell Code: Java tries to read the Stream with the defaultCharset or UTF-8 but both wont work due to the mismatch!
Instead, we just need to set the Encoding that ProcessBuilder/Powershell uses to the defaultCharset.
See the code of this commit how it can be done.